move YamlOrchestrationConfiguration to internal package
