+++ sharding-jdbc/sharding-jdbc-orchestration/src/main/java/io/shardingsphere/shardingjdbc/orchestration/internal/yaml/YamlOrchestrationMasterSlaveRuleConfiguration.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.shardingjdbc.orchestration.internal.yaml;
 
 import io.shardingsphere.core.yaml.masterslave.YamlMasterSlaveConfiguration;
-import io.shardingsphere.orchestration.yaml.YamlOrchestrationConfiguration;
+import io.shardingsphere.orchestration.internal.yaml.YamlOrchestrationConfiguration;
 import lombok.Getter;
 import lombok.Setter;
