+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/config/yaml/ProxyYamlServerConfiguration.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.shardingproxy.config.yaml;
 
 import io.shardingsphere.core.rule.Authentication;
-import io.shardingsphere.orchestration.yaml.YamlOrchestrationConfiguration;
+import io.shardingsphere.orchestration.internal.yaml.YamlOrchestrationConfiguration;
 import lombok.Getter;
 import lombok.Setter;
