+++ sharding-proxy/src/test/java/io/shardingsphere/shardingproxy/config/ShardingConfigurationLoaderTest.java
@@ -20,7 +20,7 @@ package io.shardingsphere.shardingproxy.config;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.core.yaml.masterslave.YamlMasterSlaveRuleConfiguration;
 import io.shardingsphere.core.yaml.sharding.YamlShardingRuleConfiguration;
-import io.shardingsphere.orchestration.yaml.YamlOrchestrationConfiguration;
+import io.shardingsphere.orchestration.internal.yaml.YamlOrchestrationConfiguration;
 import io.shardingsphere.shardingproxy.config.yaml.ProxyYamlRuleConfiguration;
 import org.junit.Test;
